[FIX] Remove constexpr from gap_decorator functions.
Since the gap_decorator contains a set which cannot be constexpr
constructed, the gap_decorator itslef can also not be constexpr
constructed. If it cannot be constructed, none of the member functions
can be constexpr. And since the gap_decorator_iterator stores a pointer
to the non-constexpr gap_decorator it cannot be constexpr itself.